+++ example/Example/App.h
@@ -1,0 +1,22 @@
+#pragma once
+
+struct SDL_Window;
+
+class App
+{
+public:
+	App();
+	~App();
+
+	void Init();
+	void CleanUp();
+
+	void Update();
+
+	bool quit = false;
+
+private:
+	SDL_Window* window = nullptr;
+	SDL_GLContext context = nullptr;
+};
+